Upd. WAF. Added waf for admin area.
